Decided that adding every item to a set to check whether a region of the screen needs to be redrawn is slower than simply checking each position of the screen whether an item is there

+++ RobotComplex/RecipePrototype.h
@@ -13,7 +13,7 @@ class RecipeProto {
 public:
 	std::vector<RecipeCompProto> recipe;
 	uint8_t recipeWidth;  // Determines the shape of the recipe
-	uint8_t craftingTime; // Crafting time in seconds of the recipe
+	float craftingTime; // Crafting time in seconds of the recipe
 	std::string animation;// Animation that plays when crafting
 	RecipeProto(std::vector<RecipeCompProto> recipe, uint8_t recipeWidth, uint8_t craftingTime, std::string animation)
 	{
